+++ noodles-cram/src/async/reader/data_container/compression_header/preservation_map.rs
@@ -1,4 +1,4 @@
-use noodles_sam as sam;
+use noodles_sam::record::data::field::{value::Type, Tag};
 use tokio::io::{self, AsyncBufRead, AsyncBufReadExt, AsyncRead, AsyncReadExt};
 
 use crate::{
@@ -122,9 +122,7 @@ async fn read_raw_tag_ids_dictionary<R>(reader: &mut R) -> io::Result<Vec<Vec<re
 where
     R: AsyncBufRead + Unpin,
 {
-    use sam::record::data::field::value::Type;
-
-    use self::record::tag::Key;
+    use crate::record::tag::Key;
 
     const NUL: u8 = 0x00;
 
@@ -145,12 +143,13 @@ where
         for chunk in keys_buf.chunks_exact(3) {
             let (t0, t1, ty) = (chunk[0], chunk[1], chunk[2]);
 
-            let tag = [t0, t1];
+            let tag = Tag::try_from([t0, t1])
+                .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
+
             let ty =
                 Type::try_from(ty).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
 
             let key = Key::new(tag, ty);
-
             line.push(key);
         }
 
@@ -224,8 +223,6 @@ mod tests {
 
     #[tokio::test]
     async fn test_read_tag_ids_dictionary() -> io::Result<()> {
-        use sam::record::data::field::value::Type;
-
         use self::record::tag::Key;
 
         let data = [
@@ -242,10 +239,10 @@ mod tests {
 
         let expected = TagIdsDictionary::from(vec![
             vec![
-                Key::new([b'C', b'O'], Type::String),
-                Key::new([b'N', b'H'], Type::UInt8),
+                Key::new(Tag::Comment, Type::String),
+                Key::new(Tag::AlignmentHitCount, Type::UInt8),
             ],
-            vec![Key::new([b'P', b'G'], Type::String)],
+            vec![Key::new(Tag::Program, Type::String)],
         ]);
 
         assert_eq!(actual, expected);